AllowExportFiles & SMS Templates on Action Rules into their own components
