+++ pom.xml
@@ -16,8 +16,11 @@
 
   <properties>
     <maven.compiler.release>17</maven.compiler.release>
+    <maven.compiler.source>17</maven.compiler.source>
+    <maven.compiler.target>17</maven.compiler.target>
     <log4j2.version>2.17.0</log4j2.version>
   </properties>
+
 
   <dependencyManagement>
     <dependencies>
