+++ src/main/resources/application.yml
@@ -77,7 +77,7 @@ notifyservice:
 
 iapaudience: DUMMY
 
-dummyuseridentity-allowed: false # This ** MUST ALWAYS!! ** be false in production!!!
+dummyuseridentity-allowed: true # This ** MUST ALWAYS!! ** be false in production!!!
 dummyuseridentity: dummy@fake-email.com
 dummysuperuseridentity: dummy@fake-email.com
